+++ packages/playground/src/sidebar/fixtures/npmPlugins.ts
@@ -34,7 +34,7 @@ export const allNPMPlugins = [
     description:
       "TypeScript playground plugin to display instant documentation.",
     author: "atanasster",
-    href: "https://github.com/ccontrols/structured-types",
+    href: "https://www.npmjs.com/package/%40structured-types%2Finstant-documentation-plugin",
   },
   {
     name: "Code Show Flow",
@@ -113,18 +113,18 @@ export const allNPMPlugins = [
     href: "https://github.com/orta/playground-ts-symbols",
   },
   {
-    name: "Dts Plugin",
-    id: "playground-dts-plugin",
-    description: "Generate d.ts from JS",
-    author: "sanders_n",
-    href: "https://github.com/sandersn/playground-dts-plugin",
-  },
-  {
     name: "GitHub Explorer",
     id: "ts-playground-plugin-github-explorer",
     description: "Explore a GitHub repo on TypeScript playground",
     author: "emnudge",
     href: "https://www.npmjs.com/package/ts-playground-plugin-github-explorer",
+  },
+  {
+    name: "Dts Plugin",
+    id: "playground-dts-plugin",
+    description: "Generate d.ts from JS",
+    author: "sanders_n",
+    href: "https://github.com/sandersn/playground-dts-plugin",
   },
   {
     name: "TypeScript Json Schema",