+++ packages/ts-twoslasher/CHANGELOG.md
@@ -1,3 +1,5 @@
+# @typescript/twoslash
+
 ## 3.1.0
 
 Adds support for `// ---cut-after---`, e.g.